feat: created usedisablezoom hook and updated pages
