+++ components/UI/FooterBanner/FooterBanner.tsx
@@ -1,10 +1,12 @@
 import { Logo } from "@/components/UI/Button/Logo";
 import RenderIf from "@/components/UI/Common/RenderIf";
+import useDisableZoom from "@/hooks/useDisableZoom";
 import HighlightOffIcon from "@mui/icons-material/HighlightOff";
 import { useEffect, useState } from "react";
 import styles from "./FooterBanner.module.css";
 
 export default function FooterBanner() {
+  useDisableZoom();
   const [hideFooter, setHideFooter] = useState(true);
 
   const handleOffIconClick = () => {
